Solved 5 problems from CF educational codeforces round 21 and a geometry problem from cs academy

+++ CS Academy/Round #32/Square Root Frac (Easy).cpp
@@ -54,7 +54,7 @@ int main()
             double error = abs(round(check) - check );
             if ( error < minError )
             {
-                ans = round(check);
+                ans = (int)round(check);
                 minError = error;
             }
 //            if ( abs(round(check) - check ) < SM )
